DIS-58 Refactor repository namespace
