+++ code/src/main/java/DispoSell/repositories/FurnitureCategoryRepository.java
@@ -1,4 +1,4 @@
-package DispoSell.repository;
+package DispoSell.repositories;
 
 import DispoSell.models.FurnitureCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
