+++ code/src/main/java/DispoSell/repositories/FurnitureConditionRepository.java
@@ -1,4 +1,4 @@
-package DispoSell.repository;
+package DispoSell.repositories;
 
 import DispoSell.models.FurnitureCondition;
 import org.springframework.data.jpa.repository.JpaRepository;
