+++ code/src/main/java/DispoSell/repositories/FurnitureMediaRepository.java
@@ -1,6 +1,5 @@
-package DispoSell.repository;
+package DispoSell.repositories;
 
-import DispoSell.models.Furniture;
 import DispoSell.models.FurnitureMedia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
