+++ code/src/main/java/DispoSell/repositories/FurnitureRepository.java
@@ -1,7 +1,6 @@
-package DispoSell.repository;
+package DispoSell.repositories;
 
 import DispoSell.models.Furniture;
-import DispoSell.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
