+++ code/src/main/java/DispoSell/security/services/UserDetailsServiceImpl.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import DispoSell.models.User;
-import DispoSell.repository.UserRepository;
+import DispoSell.repositories.UserRepository;
 
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
